+++ CHANGES.rst
@@ -2,8 +2,23 @@ Changelog
 =========
 
 
-1.0a1 (unreleased)
+1.0.3
+----------------
+- Added masonry and other layouts from [txbizlaw.customviews] tiles to collection summary tile
+  [b4oshany]
+
+
+1.0.1 - 1.0.2
+------------------
+- Improve Section Content
+  [b4oshany]
+- Add Display mode to mosaic content and collection summary tile form
+  [b4oshany]
+
+1.0a1
 ------------------
 
 - Initial release.
   [alecm]
+
+[txbizlaw.customviews]: https://gitlab.com/alteroo/texasbusinesslaw/txbizlaw.customviews